Import vmxfile class.
